requirements.txt
streamlit
google-generativeai
chromadb
pypdf
langchain
langchain-experimental
langchain_google_genai
toml
texify
scikit-image
